ragas_eval/pyproject.toml
[project]
name = "ragas-eval"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-adk",
    "google-genai",
    "ragas",
    "arize-phoenix",
]
